@@ -35,10 +35,12 @@ namespace ProjetoWEBMVC_CSharp_Nelio
             });
 
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            //services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
 
             services.AddDbContext<ProjetoWEBMVC_CSharp_NelioContext>(options =>
                     options.UseSqlite(Configuration.GetConnectionString("ProjetoWEBMVC_CSharp_NelioContext")));
+            services.AddControllersWithViews();
+
 
             services.AddScoped<SeedingService>();
             services.AddScoped<SellerService>();
@@ -60,13 +62,16 @@ namespace ProjetoWEBMVC_CSharp_Nelio
 
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseCookiePolicy();
 
-            app.UseMvc(routes =>
+            app.UseAuthentication();
+            app.UseRouting();
+
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
